Add files via upload
This is a simplified, brief script which 
(1) Randomly generates gaussian data,
(2)Defines and trains the single hidden layer neural network 
(3) trains the model to predict the gaussians for 100 species using their eigenvalues as the input, and 
(4) writes the results to a csv file accessible on the user's device, and which can be used to graph results separately if they wish to experiment with the code
e.g. the epochs, learning rate, Gaussian widths, etc.

Doing a plot of  say plt.plot(results['x'][0], results['y'][0]) using results.csv will graph the approximation of y = e**(x-eigenvalue)**2/(2(sigma)**2).